Merge

+++ tasks/views.py
@@ -47,8 +47,11 @@ def task_info(request, task_id):
 
 @login_required(login_url='/')
 def task_del(request, task_id):
-    try:
-        TaskModel.objects.get(id=task_id).delete()
-    except ObjectDoesNotExist:
-        pass
-    return redirect("/tasks")
+    user = UserModel.objects.get(id=id)
+    permission = user.job_title
+    if permission:
+        try:
+            TaskModel.objects.get(id=task_id).delete() # вношу изменения в этом декораторе
+        except ObjectDoesNotExist:
+            pass
+        return redirect("/tasks")
